change file name of check_reservation.dart
